+++ darglint/docstring/sphinx.py
@@ -28,6 +28,9 @@ from ..lex import (
     lex,
     condense,
 )
+from ..config import (
+    Strictness,
+)
 
 
 class Docstring(BaseDocstring):
@@ -302,3 +305,32 @@ class Docstring(BaseDocstring):
         # TODO: Implement me!
         t = []  # type: List[Tuple[Callable, Tuple[int, int]]]
         return (x for x in t)
+
+    def satisfies_strictness(self, strictness):
+        # type: (Strictness) -> bool
+        """Return true if the docstring has no more than the min strictness.
+
+        Args:
+            strictness: The minimum amount of strictness which should
+                be present in the docstring.
+
+        Returns:
+            True if there is no more than the minimum amount of strictness.
+
+        """
+        sections = {
+            section for section in Sections
+            if self.get_section(section)
+        }
+        if strictness == Strictness.SHORT_DESCRIPTION:
+            return sections == {Sections.SHORT_DESCRIPTION}
+        elif strictness == Strictness.LONG_DESCRIPTION:
+            return sections in [
+                {Sections.SHORT_DESCRIPTION},
+                # Shouldn't be possible, but if it is in the future, then
+                # we should allow this.
+                {Sections.LONG_DESCRIPTION},
+                {Sections.SHORT_DESCRIPTION, Sections.LONG_DESCRIPTION},
+            ]
+        else:
+            return False