+++ darglint/integrity_checker.py
@@ -43,6 +43,7 @@ from .error_report import (
 )
 from .config import (
     Configuration,
+    Strictness,
 )
 
 
@@ -61,7 +62,8 @@ class IntegrityChecker(object):
                  config=Configuration(
                      ignore=[],
                      message_template=None,
-                     style=DocstringStyle.GOOGLE
+                     style=DocstringStyle.GOOGLE,
+                     strictness=Strictness.FULL_DESCRIPTION,
                  ),
                  raise_errors=False
                  ):
@@ -102,6 +104,11 @@ class IntegrityChecker(object):
                     function.docstring,
                 )
                 self._check_variables()
+            if self.config.strictness != Strictness.FULL_DESCRIPTION:
+                if self.docstring.satisfies_strictness(
+                    self.config.strictness
+                ):
+                    return
             if self.docstring.ignore_all:
                 return
             self._check_parameters()
@@ -176,7 +183,7 @@ class IntegrityChecker(object):
         fun_type = self.function.return_type
         doc_type = self.docstring.get_types(Sections.RETURNS_SECTION)
         if not doc_type or isinstance(doc_type, list):
-            doc_type = ''
+            doc_type = None
         if fun_type is not None and doc_type is not None:
             if fun_type != doc_type:
                 if self.config.style == DocstringStyle.GOOGLE: